weather: add all four weather presets on per-location basis
Expand the AMK dynamic weather script to be able to select one of the
four NightVz weather presets on per-location basis: September, October,
Depressive, and Mist.
They are selected in the @dynweather_types array at the beginning of
amk_mod.script. It works the way that at the initial loading on a
particular location, weather manager switches to the new preset
completely, including weather probabilities. Static weather locations,
such as undergrounds and Yantar, also have one of the presets, but only
to generate dynamic news and variables.
All four are sprinkled across 12 outdoor locations by 3 loc per preset.
Yantar weather is a bit unique: it is static, but the static section
gets picked depending on the preset set in the abovementioned array. So,
it is possible to switch preset for this location the same way as for
the "standard" locations.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>