artefact/receipts: restore Electric Dikobraz and Giant's Small Brother
There are 2 receipts lost in the original AMK release. In WAWKA, they
were restored as a part of the reward for one scientist's quest. It
already contains receipt as a reward, so it doesn't look like this was
the original intention.
Instead, use the more common patch, which brings those two as
Osvedomitel's info available for buying. AMK even included dialogs for
that originally, but they got omitted in the release for some reason.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>