gameplay/balance: fix imbalance from Rangs' rewards
Rangs mod's rewards are overkill. Nobody would give stuff like elite
detector or exo or Gauss to someone just because that someone is good
at something. Such rewards break the NLC's balance hardly.
Make them just symbolic, but sometimes handy. It can be a box of AK
ammo or a medikit. Only the last, 7th grade is rewarded greatly, this
includes a good (but not the best) suit or a powerful weapon. 7th grade
is only accessible near the game's end or even in the freeplay.
With this change, Rangs mod introduces more interesting difficulties
than rewards, such as: lower probabilities of successful artifact
transmutation, less tasks available at the start, higher prices etc.
If this would cause imbalance in the opposite direction, it will be
fixed.
Additionally, make trader's rangs symmetric to the rest: let there be
7 of them, not 6.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>