gameplay/rangs: make its logics more consistent
Rangs mod is not consistent in terms "bonuses vs obligatories" when it
comes to the stuff rangs affect. For example, in order to be able to
brew artifacts in anomalies, you need to reach certain rang. But at the
same time, rangs affect the probability of that an NPC will share
treasure info with you only in one way -- e.g. each subsequent rang just
increases the prob. There's no such thing that would decrease this prob
if you haven't reach certain rang.
So, fix that for two aspects: discounts provided by Wawka / Syak / AMK
and the actual probability of NPCs sharing info with you. Now the
initial prob is only 10%, but then starts increasing with each new rang.
To achieve the "stock" level, you need to reach 4th rang -- around 45%
(default is 50). Same for discounts. At the beginning of the game,
prices will be x1.2 from the initial ones, but then decreasing with each
rang (PDAs found). To reach the initial x1 level, you need to obtain 3th
rang.
This makes Rangs changes way more consistent. They will only bring you
difficulties when the game starts, but then slowly start improving them,
so that on 3-4 rang you'll reach the "stock" values and then start
getting benefits: discounts, info and so on.

As a bonus, change the confusing function names in sak_dialog. The
corresponding dialog actually requires 50000 RUB, but the functions have
"40000" in their names.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>