scripts/hit_effector: properly implement effects on actor hit
Instead of checking for a hit on actor update, wire up next three
callbacks:
* on actor hit
* hit effector trigger
* before [actor] hit

and use them to implement AMK hit effector.
Effector trigger is not used for now, but can come handy later. It
allows to differenciate between different hit types: NPC (0),
monster (1), or generic (2, custom via before_hit).
From AMK's original script, the only change for now is that the
probability of dropping the weapon is now not fixed (0.2), but depends
on the hit amount (amount / 2), i.e. will be the same 0.2 on hit 0.4
and so on.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>