scripts: introduce global actor (db_actor)
In preparation of merging ZRP optimizations and to speed-up game in
general, add global _G.db_actor, which will be equal to db.actor.
Also add global db_actor_id, which is always 0 (but update it when
the actor spawns just in case).
Update all the scripts with my modifications to use new globals.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>