scripts/biodetector: replace monster checks with O(1) lookup
Current dots drawing / color picking code of biodetector is...
horrendous. 16 if-elses preceeding with 16 global var calls, which then
get passed to the function as 16 arguments. To be checked only under
certain conditions.
Now that we have separate clsid for rat, replace this with O(1) lookup.
Create 3 tables: one with the corresponding monster parts, another one
with "good" colors -- when we have such part on the belt, and the third
one with "bad" colors -- when we don't have any (used only for dogs, the
rest isn't drawn then). When we call to get_color() and hit the
condition that it's a monster, first check for the corresponding entry
in @parts by the monster's clsid, then pick one of choices from either
@good or @bad (or no color).
get_color() gets called on each actor update callback call, so this is
a great optimization when a bioradar is on the belt. Even when it's not,
previously the update function was checking for 16 conditions before
actually checking if we have a bio at all.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>