scripts/fenechka: fix spoiling food from radiation
When I integrated Fenechka by Tag, I added randomness to not spoil
everything in the inventory at once. However, the condition check is
called once per tick (1 ms), so eventually after a bunch of ticks all
food would become radioactive.
To fix that, add a flag, which reflects whether the code was run
already. IOW, once we enter a zone with radiation level > 0.4, food gets
randomly spoiled and then the flag gets set, which prevents the code
from running again, until we leave the zone and enter it (or any other)
again. So now it can spoil only once per radioactive spot, but player
needs to be careful, as it can walk at its edge and enter/leave
the zone again and again, making the flag being set and dropped
continously.
To not spoil food on every save/load in a radioactive spot, write this
flag to savegame. It can be helpful later on if any other code depending
on the actor being in such zone is added.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>