scripts/inventory: improve belt items accounting
X-Ray extensions allow to greatly improve and optimize inventory.script,
which is used to track belt items. It is needed to check whether the
actor has monster parts on the belt when using biodetector and will be
also used for ammo and grenades on belt scripts.
Instead of inserting a fake item (separator) and then track stuff in two
ticks (1 -- inserting and iterating, 2 -- removing and calling for
updates), use new callbacks: on_item_belt(), on_item_ruck(),
on_item_slot(). Just record items by their IDs on triggering the first
callback and remove their records on the second and third. The update
callback is then not needed anymore, same for separator item.
The idea belongs to Ray Twitty, but his sa_callbacks.script is present
in this mod anyway with all the copyrights :p

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>