scripts/artefacts: add artefacts ranking and stats accounting
The original SoC scripts completely miss artefact accounting for the
player. Ideally, you should get some rank credits for each artefact,
just how it's done for NPCs, but this doesn't happen.
To address this, pick SA's logics, which accounts found artefacts. It
introduces new field for se_artefact, which is false until you take this
artefact. If you take it for the first time, you get credited for this
art according to the reward table.
The table itself differs a bit from the original Twitty's code. Instead
of specifying scores per each artefact manually, just rank them by their
cost. This is a bit cheap, but in general reliable, as they only have
several different cost values available: 1000 (1st level), 2500, 5000
and then it starts to be a bit more diverse: 8000, 10000, 12k, 20, 50
and so on. So divide them in 5 levels, the first three would +/- reflect
the original game's art level division. After the third level (5000),
just define two more: 10k and 20k, to account artmods and uniques.

A couple implementation details specific to this mod:
* ARC containers need to spawn arts with this flag set already.
  Otherwise you'll be credited each time you take an artefact from one
  of your containers (they get spawned each time, so no flags carried
  over);
* we need to exclude Biodetector from the accounting, as it has clsid of
  artefact unfortunately (some original impl. details, no idea who and
  why). Do that for both SA's accounting and Rangs mod accounting.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>