scripts: ammo and grenades from belt (tadaaah)
Suffer lyarvo!
Port heavily modified ammo and grenades from belt (Charsi et al.).
"Heavily modified" because thanks to X-Ray exts' callbacks, there's no
more need to do tick-based updates and complicated conversions in
iterations over the whole inventory.
The logics for both ammo and grenades works like that: if some ammo (or
grenade) hits the inventory and no trade/loot/inventory UI windows are
open, convert them to "fake": same ammo, but with different section, so
the weapon can't shoot with it. If, however, the ammo is placed on the
belt, it doesn't get converted, so that it can be used for
shooting/reloading.
A bit more complicated with grenades: you can't throw them from the
belt. So, in addition to that you have "real" grenades on the belt and
"fake" ones in the inventory, for each on-belt grenade one more "fake"
gets spawned, but with the class which allows throwing it as a grenade.
They get spawned to the inventory, so you actually can throw them.
If any of the mentioned windows is open, all ammo and grenades are
"real", so you can trade and loot them. Right after the windows gets
closed, they will be converted.
This doesn't include belt slots extension -- to be addressed in the next
commit. OTOH it includes one more "suffer lyarvo": now any UI window
causes weapon holstering, not only talking. So each time you open/close
the inventory, you'd need to lock and load your weapon again (not a bug
or W/A, purely intended).

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>