ui/inventory: increase belt slots to 14 (with artefact limit)
Now that ammo needs to be placed on the belt in order to use it, 6 belt
slots are not enough (would be too harsh even for "suffer lyarvo"). FRM
mod has 12 slots, but that would mean only 3 boxes of ammo, while you
also need to use grenades, rifle launcher grenades etc...
So, extend belt slots to 14. Now 3 main slots are nicely placed in the
inventory: "small" weapon slot is 5x2, "big" weapon slot is 6x2 and belt
is 7x2, he-he. Should be enough for most cases, when not, then suffer
lyarvo.
Additionally, to prevent belt slots from abusing with artefacts, port
SA's artefacts on belt limit. It allows to specify maximum number of
arts that could be placed on the belt. When the limit is reached, the
rest of artefacts in the inventory gets "disabled", meaning you can't
place them on the belt anymore. Nice.
Use the default SA's value of 5. It correlates nicely with my own
calculations: originally, AMK/NLC/WAWKA have 6 slots: 5 for arts and 1
for a detector. We added 8 more, but purely for ammo, so the limit of
5 arts doesn't alter the balance by any means.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>