scripts: pick improvements from Solyanka w/DMX 1.3.5
There was a moment when Solyanka simply refused to work due to the game
and engine being overloaded with stuff. At that point, some developers
decided to try to optimize custom scripts to reduce processing time.
This included some AMK, NLC, NS, Kostya scripts, and most of related
changes can be ported to this build, so do this. Be careful though,
as it's better to not optimize than to break at all, this should be
improvements, not regressions. I already wrote that custom scripts
are mostly garbage (not that original GSC's aren't, but...), so this
at least makes things less ugly.

Apart from the mentioned, there's also:
* rewritten inventory limit, by me;
* safe place indicator, by me;
* rewritten statistics (no arcade score points), by me;
* some more improvements from ZRP;
* Elite detector now shows Witch's Galantine (new clsid for them),
  Johnson;
* biodetector now has new Aliens-like look;
* diary notification now shows everything, not only registered types;
* general random improvements.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>