scripts/xr_statistic: fix kills accounting
Another regression in the previous optimization round. You can't do
multi-assignments with testing values for nil in-place. Just let it fall
through to the function call, which will check for @npc anyways -- it's
an exception path, which usually should not occur at all, but if it
does, optimization doesn't matter.
Also another problem is that AMK's code were returing just `nil` instead
of `nil, nil`. I think it's not a real issue, since nothing means nil,
but fix anyways.

Fixes: d2dd40371e30 ("scripts: pick improvements from Solyanka w/DMX 1.3.5")
Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>