gameplay: introduce finite bolts
Suffer lyarvo.
Pull finite bolts from DMX 1.3.5, with some code from old Charsi's
implementation. The name is self-descriptive: bolts are now finite.
Waste and throw wisely. They can be bought from Barman, Kulinar, and
Sidorovich and also found in dead bodies (stalkers only and only up
to Radar).
When thrown, original "true" bolt gets replaced with a "fake" one in
3 secs. This is _almost_ seamless, as much as it can be. Otherwise,
it's almost impossible to pick original bolt: you should aim precisely
at it, no hints and "take distance" as with other items in the game.
Hard to do on asphalt and unreal and sicking off in grass, while "fake"
bolts is as easy to pick up as usual item.
The main difference from the original code is that now you can "restock"
your slot with a bolt _from the belt only_. This is done for better
consistence with ammo on belt and also for more suffer lyarvo. The only
exception is when your slot is empty and you pick up a bolt -- then it
hits the slot. Otherwise, it would be too much even for this build :D
Another nice to have is that they get grouped on the belt by 5. IRL, it
feels natural to be able to put 5 bolts in one pocket. 1 is too wasteful
both IRL and in the game. This is done via yet another fake item, which
treats bolts as ammo. They exist only on belts and get replaced with
"real" "fake" bolts in the ruck.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>