config/misc: pick some more ZRP fixes
Nothing special, just some typos and infop logic shortcuts. The sole
thing to mention is that now sawn-offs (BM-16 and TOZ) have
ef_weapon_type = 5 instead of 7. 7 is for long rifles. 5 prevents NPCs
to prefer ARs instead of sawn-offs when available (and AI additions
are not active, which is true for now).

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>