scripts/treasure_item: allow taking own empty treasure rucks to inventory
AMK introduced treasure_item, a ruck which you can use to create your
own treasure. However, once used, it can't be taken back, meaning it
will always stay at one place, even if you don't plan to visit it
anymore.
This is a bit of a waste, so let's adopt small piece of code from
DMX (and later OP-1 and up to the latest 2.2 now): when you grab the
last item in the ruck, it will be moved to your inventory, so that you
could use it later in some other place. But, unlike OP, which allows
taking Collector's, Kostya's and some other rucks, don't do that and
allow taking only own treasures. Otherwise, it hurts balance as you
wouldn't need to buy them anymore.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>