scripts/amk_mod: fix undefined reference to news_main.isIsolatedLevel()
When I was porting some optimizations from NS 2010, I removed the
function, which was simply returning true if the level is indoor,
but it was managing its own copy of the level list, and replaced it
simply with level.indoor(), which does that in centralized way (via
game_maps_single.ltx). However, I missed that there are calls to this
function in amk_mod.script.
Replace them with the same level.indoor().

Fixes: d2dd40371e30 ("scripts: pick improvements from Solyanka w/DMX 1.3.5")
Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>