config/artefacts/amk: anti-cheatify Bulyzhnik
WAWKA did something horrendous. -50 for radiation to the artefact
effectively dummy in the original AMK/NLC? Nah, doesn't work. Moreover,
he did that despite Syak's cheesy artmods with 0% of rejections.
Bulyzhniks will be useful later on for Kostya when Master addon is here.
But anyway, I don't want it to be so powerful, esp. given that one of
the earlier Collector treasures contains one.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>