scripts/artefact: add hit when taking artefact
This was initially developed as part of Collector. The same idea was in
NLC 6 as well, so who to credit, no idea. Anyway.
Artefacts will hit you in the wild.
Any artefact found "on the ground", i.e. not in a stash or not received
via dialog etc., will hit the actor. The actual amount and nature of hit
depends on the artefact and its class/rank.
If you have a scientific outfit put on, it will soften the hits. Unlike
Collector / OP, there are NO gloves to "safely" pick artefacts. Suffer
LYARVO.
Changes from Collector / OP:
* it hits only once -- after it's credited in the statistics, it won't.
  Like it already got "discharged". Same for stashes, dialogs etc.:
  those are already discharged. So the flag which SA introduced is used;
* after the hit, the actor drops it. Then you can pick it for the second
  time, it will be okay (similar to NLC 6);
* fixed the original Collector bug when you could take arts through open
  inventory. This is fixed in OP, but still (my fix is a bit different);
* reused some fancy OP-2.2 functions, which weren't used here.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>