gameplay/treasures: add Collector treasures from OP-2.2
Fool's paradise. Or idiot's dream. It's never too many of treasures?
I wasn't sure whether this would break the game's balance or not, but
you know... It's impossible to build the balance in such a way that
the player will always be in a shortage. There's always eventually be
too many of things, money, and so on. At least NLC 5 / NS 3 based ones.
This adds 37 treasures in total. 16 of them is from the original
Collector 0.5.5 / OP-1, 21 are new. OP-2.2 contains 54 treasures, here
I did cut out the levels not present in my mod.
Treasure contents is also different. I don't have 1488 weapons and ammo,
neither grenades etc., so it's mostly some common things, plus
non-cheaty weapons (Kriss, Saiga), plus some unique ammo (for crossbow,
P-90, AWM, and minigun). Minigun ammo is added to start adding it to
stashes, as I want to add minigun from builds later (as a unique item).
Artefacts are unchanged and not specified anywhere for trading, so you
can sell them for 0.7 of the price. I'll do some restrictions later on
to not spawn tons of money from early treasures. Maybe even add a task
for searching for the Collector himself, will see (still not sure). I do
that solely for jumping and platforming stuff anyway.
Ah right, now the actor crouches a tiny bit lower than before. This is
absolutely required in order to be able to reach the stash under the
Escape bridge. I didn't touch actor's jump height, it seems like at
least half of the treasures can be reached with the stock 6. If I'll
need more, I'll just use X-Ray Extension's feature and bind it on
something... Will see.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>