scripts: recursively replace '*actor.id*' with 0
Actor's ID is static and can't be anything but 0. actor:id() and even
@db_actor_id make no sense, although global @db_actor_id was much faster
than calling to the method each time.
Anyway, just use sed to replace them with 0 everywhere. The most profit
will be for the non-vanilla scripts from AMK, NLC, NS, and so on, as
they didn't care about optimization and were always calling to :id().

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>