textures/act: fix/restore novice and bandit stock skins
In Wawka, stock green stalker and bandit skins were replaced: the
bandit one with band01 from novices_bandits_veterans and the green
ones with two other ones already present in the pack. This was giving
the impression that there are to few unique skins, e.g. novices all
were brown-black-grey, and so on.
Turned out the original Wawka didn't have that issue, it came later
with one of the texture addons. The same addon caused military exo and
"VDV" outfit to not have any actor icon at all (just empty space).

Just revert those to the original novices_bandits_veterans, as the
pre-addon Wawka and NS 26.04.09 have, including inventory icons, NPC
icons, and actor icons.
Then, while we are here, replace novice_01 and 02 with Shadows Addon's
ones. They come from the early builds and look good, especially given
that they are now only 2 of 11, not 2 of 2 as in SA. The rest remains
the same. We're merging SA soon either way, why waste time.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>