scripts/gulag_escape: fix Lua issues in two places
This goes from the original game, but looks like only in RMA it was
fixed.
Lua is not C and should not have a ',' between the last table element
and the enclosing brace. Two Gulag jobs for the Escape novice camp had
them and I didn't even noticed that when rewriting Gulag scripts T.T
Pick this little fix not waiting for merging RMA/SA Escape changes.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>