config/gameplay: fix XML errors reported by script parser
That script parser from the PDA talk bundle is genius in its
simplicity: its naive logics uncovered several XML issues that
xrEngine forgives.
The first one is incorrect comment block in amk_characters (unused
in general, but still included) -- it misses '!'.
The second is in Syak's tasks descriptions. He left '>' in the hints
description (which point to Doctor). It could be technically correct,
but it's not correct from the grammar PoV :>
Fix both.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>