scripts/parse_xml: add option to recursively parse XML with all includes
Currently, when parse_xml finds an include in the XML, it only adds it
to the special table ("includes"), not descending there to parse it as
well.
In order to be able to parse game_tasks.xml, which in fact is a bunch of
includes (apart from the crap from NS3 haha), we need to teach it to
parse files recursively. Add such option and do the following: if the
code finds an include, read the file and concatenate the result to the
source string, replacing the include pragma itself, and return to the
beginning of the parsing routine to "try again".

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>