scripts/smart_terrain: fix population_comed == 0 for monster gulags
When I replaced comparing obj:clsid() with clsid_script_stalker in
:get_obj_community() with obj:is_stalker(), I didn't notice that
clsid gets assigned to the variable to be able to reuse it once again
if obj is a monster.
As a result, the function was returning monster_classes[nil] (== nil)
for any monster. Practically, that resulted in that population_comed
for any monster gulag was always 0 -- it's strange and funny that that
was the only symptom, the rest was working.
Since we're taking :clsid() anyway, compare directly against
clsid_script_stalker just like before -- it will be a bit quickier.
To prevent that from happening again, add an abort in case we get nil.
It doesn't trigger on the current code, which means traders
(clsid.script_trader) don't belong to any smart terrain and we
shoulnd't care about them here.

Fixes: 0342368f6c40 ("treewide: snapshot @ 08.07.23")
Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>