SA, misc: remove sa_utils / _G.round()
This function is not from SA but from the original _g.script. It was
removed in OP-2.2 in favor of math.random, which rounds the value via
string.format().
There's currently no need for the latter to be added in the build and
sa_hud_interface_mgr.script needs it only to perform round-closest
after a division. Open-code the expression and remove the function
from sa_utils.script.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>