death_manager: get rid of manual dependency table
In fact, we don't need that big dependency table.
Only ammo has dependencies.
We need to spawn ammo only for the weapons that the deadman had.

That is only 1-2 sections per one dead, which could easily be read when
processing the loot, especially if we cache the result in a local table.
Don't build dep table in advance. Store ammo_class for each weapon
section in a local table to not read the same section again and again
when we're killing 100 soldiers with the same AK.
The code piece which spawns items with dependencies can be simplified,
since we know for sure only ammo has dependencies. Just always use
create_ammo().

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>