biodetector (biomap, bioradar): improve the logics and UI a bit
Take the texture from the NS2011. It differs from the current by a bit
better quality and also by having the "adapter" by which it's "attached"
to the screen edge.
Fix biomap sweeper and fading dots drawing speed depending on the
framerate. That's why I always say we can't do anything which can be
"counted" or "measured" in update callback without calculating precise
amounts depending on the time delta since the last update.
IOW, make sweeper turn speed and dots fading speed depend on the delta.
Assume the current values are for 60 FPS and then adjust them
accordingly.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>