SA, hit effectors: don't add cam effector when the actor falls
Checking for obj:is_stalker() also triggers when the actor falls (maybe
not only), i.e. when it's not a damage caused by weapons.
Instead of making the check more and more complex, wire up
callback.on_mob_hit_to_actor (callback.hit_effector in SA) and make
the logics work the same way as it does in SA, except that we also add
simple "generic" hit effector when the monster doesn't have a special
one in its config.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>