scripts/sa_utils: add default sound to SA's send_tip()
By default, when no sound specified, the tip will be silent. It however
may be handy to play the default "pda_tip" sound, to be able to specify
only the first argument and still have the sound.
Make the function pick the sound the same way as amk.send_tip() does
that.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>