gameplay/sak: move Strelok's PDA to stash on Agroprom
NS2009 has Strelok's PDA moved from the Sarcophagus to new location.
Here we don't have any new maps and I don't plan to add any. But I liked
the idea of having it in a different place, so that you need to get
through the North of CNPP and then all the way back without the
storyline Blowout disabled on the South side, as well as earthshakes
and deadzones.
Instead, use Norman's stash from NS2011. It's located on Agroprom, below
the four tanks. To get there, you need to drop a Plasmic Caterpillar
right at the blocked entrance and the teleporter will move you there.
It's not hidden and you can see the glow from the second teleporter
inside, as well as hear Strelok's radio player there. But you won't
be able to access it without having recipe from Sakharov and without
finding the note from Strelok near the Monolith.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>