scripts: get rid of prefetch() and heavy calls on script load
According to OGSE/OGSR, prefetch() is not reliable and it's enough and
better to just do `if _G[file]`. Remove it everywhere and replace with
explicit checks.
This also allows to define early module load stages. For now, there are
two: at the beginning of _g load and then after RvP() and some variables
are initialized. The first one is used to recheck JIT and load sa_init
and xr_extensions (needed for Lua ext. etc). The second is for lua_ext
and reap_lua as they might require some pre-initialization.
Also, don't fill tables and news_manager data when _g is read. It's then
done once during the main menu initialization and then one more time
during the game loading. It's enough to do that in start_game_callback.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>