scripts/reap: reseed the random every 5 seconds
Make random a bit more random.
Also, since all of the intervals for delayed updates are compatible
(1 second after 200 ms which is x5, then 5 seconds after 1 which is
again x5), we can leave as early as we get the first one not expired
yet. This is the same as how we sort timers and exit on the first
not expired.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>