gameplay/treasures: add Strelok's treasures from Master addon
Including updates to Kostya's and all the related treasures from OP-2.x
and naxac's addon.
Sakharov now can make a new teleport prototype for the same money as for
the rest. Previously, I had functions for that, but not the dialog :D
Also, xr_effects is cleaned out of the Kostya's functions. Since AMK,
we can define infop test and set functions outside of xr_conditions and
xr_effect, let's use that and leave the original only for the generic
routines.
Also add dezombifying med, mainly to put them in the treasures, but it
will be useful later as well, as I plan to add actor zombification later
:D
There were 2 bonus treasures in Limask (or Red Forest), which I replaced
with 2 treasures on Agro, and one main on the Generators, which I
replaced with Volazar's treasure in X-16. With the change that the final
destination is the brain itself. I like that it allows to visit the
brain's top cover (added a couple decorations there), plus X-16 is not
a common place for treasures.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>