scripts/gulag_garbage: fix .check() for gar_psevdowounded
check() callback implementation may take @se_obj as a real server object
if it's a stalker or nil if it's a monster. Thus, each .check() must
check whether the param is nil.
This wasn't true for gar_psevdowounded, so that the first monster
considering joining this smart crashed the game.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>