scripts/amk_mod: get rid of blowout_type > 1
Let the CNPP have the same effects as the indoors. Quick timer respin
means no anomaly and artefact re-placement etc. Also it's just not
logical to not feel anything there.
Introduce level.blowout() instead of amk_mod.blowout_type(), which
returns true if we need to play full cycle of effects (outdoor non-CNPP
maps). This can be overriden in game_maps_single.ltx, which is how it's
implemented for the CNPP.
Also, get rid of the stateless init of sa_callbacks/sa_utils. Move its
init to __init() function and call it on Actor's :__init(). This is a
little earlier than previously (on net_spawn), which doesn't hurt.
reap_weather needs level.blowout() from sa_utils, so it must be
initialized first.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>