scripts/xr_logic: whitelist xr_conditions.actor_in_zone()
When I added a check for non-nil Actor and npc before running
infop.func(), which in theory can end up with no checks, it was working
fine, but recently I caught a crash that xr_conditions.actor_in_zone()
was called before the actor is initialized.
actor_in_zone() has all needed checks inside, so we can add it to a newly
created whitelist. I still don't believe we need actor_proxy.
Let it be just one 'if' for now. Then, if needed, we could convert it
to a hastable later on.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>