scripts/reap: fix weapon show/hide when the game loads
Recently, I caught a bug that the weapon was shown when moving it to
slot despite that it's disabled when the inventory is opened.
Turned out the timer to enable weapon show after an item is used was
triggering when the game was loading. The weapon was enabled like it
normally was, but such strange bugs appeared then.
Spin a timer if the game is still loading when wpn_show_enable() or
wpn_hide_disable() is called until it is fully loaded. The bug
disappeared and weapon is not shown anymore through the inventory.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>