meshes/weapons: overhaul weapon HUDs with SA extensions
This includes some fixes and a couple new animations (e.g. when
moving but not sprinting).
HUDs not imported from SA:
* AK-74 - HUD orientation/position were changed and I suck in
  reconfiguring it. Kept the original HUD + Normal Weapons
  Orientation;
* long BM-16 - same as above;
* SPAS-12 - Gosuke's extensions are preferred (reload etc.);
* SVD - Gosuke extensions (sprint anim);
* SVU - same as SVD;
* Wincheaster 1300 - same as SPAS-12.

For the rest, define anim_idle_moving and other non-optional in SA and
remove the checks for anim existence from its code.
Also includes some small fixes such as binoculars name, translations
etc. And a bug when binocs wasn't moved to the slot automatically, but
you couldn't do that manually as slot 4 doesn't exist in the inventory
and the game was just crashing. I'll add these slots later and then
we'll remove default_to_ruck again.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>