gameplay/weapons: the Great Weapon Overhaul
Used models and textures from:

* Build by Sam 5.2.1
* Build weapon pack
* Gunslinger WP SoC Edition
* New Arsenal 5.1
* OP 2 DSh Mod (NLC etc.)
* OP 2.2
* SoC weapon pack
* Shadows Addon 0.8.9 weapon packs (CS and NLC)
* Weapon Pack for TT2

Almost each weapon now has a second skin. For the vanilla weapons, the
first skin is the vanilla one, the second is the new one. For the
non-vanilla ones, either only one skin exists or some second good skin
from the above is taken.
All the scripts are adjusted to recognize different skins of the same
weapon.
Lots of scripting, e.g. animations processing, also support for 3D
scopes, dynamic zoom_rotate_time and scope_zoom_offset etc. etc.
More to come.

The weapon textures are now in a separate gamedata.db archive to not
bloat the repo.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>