engine: switch to OGSR 2.1.421
I finally made it.
I tried several times and each time I faced some issues which blocked
me from doing this. Sometimes it was engine bugs, but later -- just
functionality missing in OGSR in comparison with xrExtensions.
But given that OGSR is open source... Everything can be added or changed
there.
There is a couple TODOs, such as restoring the original controller aura,
but most of functionality works now. Monster attacks, weapon features,
slow medicine, cars, ...
I playtested it a bit and found only minor issues not related to the
engine switch at all. On the shadow map texture size of 4096, FPS is
good even with detail_density of 0.1 and detail_radius of 100, but now
the shadows are great on 4096, while on the original engine, less than
8192 was looking awful.
New engine gave some good features right from the start, such as talk
item accelerators, item highlight and condition bars, not speaking of
graphics. I also unblocked monster effectors in the engine instead of
scripting it (sa_main), as they have both PP and cam effects, but in
scripts, you could only add a cam effector.
This includes yet another font refactoring. Along with texture800 and
texture1600, I introduced texture_hidpi in the engine, so that 2x fonts
will be automatically picked on 4k+ resolutions without needing to
increase font scale. In addition to that, this texture pick now works
for font_console and font_hud, as without that, you'd always have messed
up font in the console due to that it doesn't get scaled.
Fonts are declared as multibyte, so starting from this commit, it's safe
to encode config/text/ in UTF-8. Not sure about config/gameplay/ and
scripts/ -- the strings must go through game.translate_string()
correctly, so I'll either way need to move all the string to text/ later
on to make everything UTF-8.
Lots of stuff to tell and explain, just try it out.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>