scripts/misc: use new engine callback to init game_graph()
Use the callback I added in the engine tree which gets called right
after the game graph is built, so that you can call game_graph() to
assign it to a global var. In OGSR, graph init happens later than
on 1.0006 or OP-2.2, so you can't call this function right away in
_g.script.
I hacked it around by initalizing it just before the first variable
access in smart_terrain.script, but it didn't look good. A callback
looks more natural here.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>