text/ui: fix missing 'ui_st_description' translation
Turns out this string was in one of the multiplayer translation files
which I stopped including in the tagged commit to not load always
unused resources.
Not the first time when translation strings find where they don't
belong. Add the missing string instead of one multiplayer line which,
in contrary, was in a single player file :clownface:

Fixes: 1f036a5db4c3 ("engine: switch to OGSR 2.1.421")
Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>