scripts/bind_actor: stop moving actor's outfit out of its slot
This was needed for 1.0006 as it tends to reset actor visual each time
a new outfit is picked up. On OGSR, this is not needed anymore as it
changes actor visuals correctly. Instead, it introduced issue when the
actor appeared without an outfit each time you pick a new one, due to
'ruck_flag_preferred' option.
Just remove this workaround which both removes unneded code and fixes
a real issue.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>