scripts/amk/news: fix more weapon type news
In the weapon overhaul commit, I introduced "news_weapon_type" field to
describe which news string to pick when someone is killed from that
weapon.
Turned out there are more places to fix, so do that now, and also move
the hardcoded strings to config/text/, where they should be.

Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>