scripts/arc: fix containers respawning after giving them away
The tagged commit started releasing objects when you give them away to
an NPC if they're not quest. This might've conflict with the scripts
which use old methods of checking that you used an item -- by listening
to on_item_drop.
In particular, ARC containers code treats release() as usage. Due to
that, when completing the task to find the Night Star, the container
Sidorovich gave you was released -> UI dialog with its content was
appearing -> clicking "Close" would respawn it back to you.
Just use the use callback and don't react on release() anyhow.

Fixes: 0342368f6c40 ("treewide: snapshot @ 08.07.23")
Signed-off-by: Alexander Lobakin <alobakin@mailbox.org>